web_requirements.txt
# Web Framework & API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
jinja2>=3.1.0
python-multipart>=0.0.6

# Job Queue & Background Tasks
redis>=5.0.0
rq>=1.15.0

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0

# Data Validation & Serialization
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Configuration
pyyaml>=6.0.1
python-dotenv>=1.0.0

# File Handling
aiofiles>=23.2.0
python-magic>=0.4.27

# AI & ML (Core)
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
diffusers>=0.21.0
transformers>=4.30.0
accelerate>=0.20.0
controlnet-aux>=0.0.6

# Computer Vision
opencv-python>=4.8.0
numpy>=1.24.0
pillow>=10.0.0
scikit-image>=0.21.0
tqdm>=4.66.0

# Face Processing
face-recognition>=1.3.0
mediapipe>=0.10.0
insightface>=0.7.0
onnxruntime>=1.16.0

# Video Processing
ffmpeg-python>=0.2.0
moviepy>=1.0.3

# Image Enhancement
gfpgan>=1.3.0
basicsr>=1.4.0
realesrgan>=0.3.0

# Utilities
requests>=2.31.0
httpx>=0.25.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
isort>=5.12.0